refactored message formatter
