+++ framework/yii/i18n/MessageFormatter.php
@@ -7,10 +7,8 @@
 
 namespace yii\i18n;
 
-if (!class_exists('MessageFormatter', false)) {
-	require_once(__DIR__ . '/FallbackMessageFormatter.php');
-}
-defined('YII_INTL_MESSAGE_FALLBACK') || define('YII_INTL_MESSAGE_FALLBACK', false);
+use yii\base\Component;
+use yii\base\NotSupportedException;
 
 /**
  * MessageFormatter enhances the message formatter class provided by PHP intl extension.
@@ -25,53 +23,133 @@ defined('YII_INTL_MESSAGE_FALLBACK') || define('YII_INTL_MESSAGE_FALLBACK', fals
  *   However it is highly recommended that you install [PHP intl extension](http://php.net/manual/en/book.intl.php) if you want
  *   to use MessageFormatter features.
  *
+ * FallbackMessageFormatter is a fallback implementation for the PHP intl MessageFormatter that is
+ * used in case PHP intl extension is not installed.
+ *
+ * Do not use this class directly. Use [[MessageFormatter]] instead, which will automatically detect
+ * installed version of PHP intl and use the fallback if it is not installed.
+ *
+ * It is highly recommended that you install [PHP intl extension](http://php.net/manual/en/book.intl.php) if you want to use
+ * MessageFormatter features.
+ *
+ * This implementation only supports to following message formats:
+ * - plural formatting for english
+ * - select format
+ * - simple parameters
+ *
+ * The pattern also does NOT support the ['apostrophe-friendly' syntax](http://www.php.net/manual/en/messageformatter.formatmessage.php).
+ *
+ * Messages that can be parsed are also not necessarily compatible with the ICU formatting method so do not expect messages that work without intl installed
+ * to work with intl
+ *
  * @author Alexander Makarov <sam@rmcreative.ru>
  * @author Carsten Brandt <mail@cebe.cc>
  * @since 2.0
  */
-class MessageFormatter extends \MessageFormatter
+class MessageFormatter extends Component
 {
-	/**
-	 * Format the message.
-	 *
-	 * @link http://php.net/manual/en/messageformatter.format.php
-	 * @param array $args Arguments to insert into the format string.
-	 * @return string|boolean The formatted string, or false if an error occurred.
-	 */
-	public function format($args)
-	{
-		if ($args === []) {
-			return $this->getPattern();
-		}
+	private $_errorCode = 0;
+	private $_errorMessage = '';
 
-		if (version_compare(PHP_VERSION, '5.5.0', '<') && !YII_INTL_MESSAGE_FALLBACK) {
-			$pattern = self::replaceNamedArguments($this->getPattern(), $args);
-			$this->setPattern($pattern);
-			$args = array_values($args);
-		}
-		return parent::format($args);
+	/**
+	 * Get the error text from the last operation
+	 * @link http://php.net/manual/en/messageformatter.geterrorcode.php
+	 * @return string Description of the last error.
+	 */
+	public function getErrorCode()
+	{
+		return $this->_errorCode;
 	}
 
 	/**
-	 * Quick format message.
+	 * Get the error text from the last operation
+	 * @link http://php.net/manual/en/messageformatter.geterrormessage.php
+	 * @return string Description of the last error.
+	 */
+	public function getErrorMessage()
+	{
+		return $this->_errorMessage;
+	}
+
+	/**
+	 * Formats a message via ICU message format.
 	 *
 	 * @link http://php.net/manual/en/messageformatter.formatmessage.php
-	 * @param string $locale The locale to use for formatting locale-dependent parts.
-	 * @param string $pattern The pattern string to insert things into.
-	 * @param array $args The array of values to insert into the format string.
-	 * @return string|boolean The formatted pattern string or false if an error occurred.
+	 * @param string $language The locale to use for formatting locale-dependent parts
+	 * @param string $message The pattern string to insert things into.
+	 * @param array $params The array of values to insert into the format string
+	 * @return string|boolean The formatted pattern string or `FALSE` if an error occurred
 	 */
-	public static function formatMessage($locale, $pattern, $args)
+	public function format($language, $message, $params)
 	{
-		if ($args === []) {
-			return $pattern;
+		$this->_errorCode = 0;
+		$this->_errorMessage = '';
+
+		if ($params === []) {
+			return $message;
 		}
 
-		if (version_compare(PHP_VERSION, '5.5.0', '<') && !YII_INTL_MESSAGE_FALLBACK) {
-			$pattern = self::replaceNamedArguments($pattern, $args);
-			$args = array_values($args);
+		if (!class_exists('MessageFormatter', false)) {
+			return $this->fallbackFormat($language, $message, $params);
 		}
-		return parent::formatMessage($locale, $pattern, $args);
+
+		if (version_compare(PHP_VERSION, '5.5.0', '<')) {
+			$message = $this->replaceNamedArguments($message, $params);
+			$params = array_values($params);
+		}
+		$formatter = new \MessageFormatter($language, $message);
+		if ($formatter === null) {
+			$this->_errorMessage = "Unable to format message in language '$language'.";
+			return false;
+		}
+		$result = $formatter->format($params);
+		if ($result === false) {
+			$this->_errorCode = $formatter->getErrorCode();
+			$this->_errorMessage = $formatter->getErrorMessage();
+			return false;
+		} else {
+			return $result;
+		}
+	}
+
+	/**
+	 * Parse input string according to a message pattern
+	 *
+	 * @link http://www.php.net/manual/en/messageformatter.parsemessage.php
+	 * @param string $language The locale to use for formatting locale-dependent parts
+	 * @param string $pattern The pattern with which to parse the message.
+	 * @param array $message The message to parse, conforming to the pattern.
+	 * @return string|boolean An array containing items extracted, or `FALSE` on error.
+	 * @throws \yii\base\NotSupportedException when PHP intl extension is not installed.
+	 */
+	public function parse($language, $pattern, $message)
+	{
+		$this->_errorCode = 0;
+		$this->_errorMessage = '';
+
+		if (!class_exists('MessageFormatter', false)) {
+			throw new NotSupportedException('You have to install PHP intl extension to use this feature.');
+		}
+
+		// TODO try to support named args
+//		if (version_compare(PHP_VERSION, '5.5.0', '<')) {
+//			$message = $this->replaceNamedArguments($message, $params);
+//			$params = array_values($params);
+//		}
+		$formatter = new \MessageFormatter($language, $pattern);
+		if ($formatter === null) {
+			$this->_errorMessage = "Unable to parse message in language '$language'.";
+			return false;
+		}
+		$result = $formatter->parse($message);
+		if ($result === false) {
+			$this->_errorCode = $formatter->getErrorCode();
+			$this->_errorMessage = $formatter->getErrorMessage();
+			return false;
+		} else {
+			return $result;
+		}
+
 	}
 
 	/**
@@ -123,5 +201,156 @@ class MessageFormatter extends \MessageFormatter
 			$d += 1 - substr_count($parts[$i], '}');
 		}
 		return $pattern;
+	}
+
+	/**
+	 * Fallback implementation for MessageFormatter::formatMessage
+	 * @param string $language The locale to use for formatting locale-dependent parts
+	 * @param string $message The pattern string to insert things into.
+	 * @param array $params The array of values to insert into the format string
+	 * @return string|boolean The formatted pattern string or `FALSE` if an error occurred
+	 */
+	protected function fallbackFormat($locale, $pattern, $args = [])
+	{
+		if (($tokens = $this->tokenizePattern($pattern)) === false) {
+			return false;
+		}
+		foreach($tokens as $i => $token) {
+			if (is_array($token)) {
+				if (($tokens[$i] = $this->parseToken($token, $args, $locale)) === false) {
+					return false;
+				}
+			}
+		}
+		return implode('', $tokens);
+	}
+
+	/**
+	 * Tokenizes a pattern by separating normal text from replaceable patterns
+	 * @param string $pattern patter to tokenize
+	 * @return array|bool array of tokens or false on failure
+	 */
+	private function tokenizePattern($pattern)
+	{
+		$depth = 1;
+		if (($start = $pos = mb_strpos($pattern, '{')) === false) {
+			return [$pattern];
+		}
+		$tokens = [mb_substr($pattern, 0, $pos)];
+		while(true) {
+			$open = mb_strpos($pattern, '{', $pos + 1);
+			$close = mb_strpos($pattern, '}', $pos + 1);
+			if ($open === false && $close === false) {
+				break;
+			}
+			if ($open === false) {
+				$open = mb_strlen($pattern);
+			}
+			if ($close > $open) {
+				$depth++;
+				$pos = $open;
+			} else {
+				$depth--;
+				$pos = $close;
+			}
+			if ($depth == 0) {
+				$tokens[] = explode(',', mb_substr($pattern, $start + 1, $pos - $start - 1), 3);
+				$start = $pos + 1;
+				$tokens[] = mb_substr($pattern, $start, $open - $start);
+				$start = $open;
+			}
+		}
+		if ($depth != 0) {
+			return false;
+		}
+		return $tokens;
+	}
+
+	/**
+	 * Parses a token
+	 * @param array $token the token to parse
+	 * @param array $args arguments to replace
+	 * @param string $locale the locale
+	 * @return bool|string parsed token or false on failure
+	 * @throws \yii\base\NotSupportedException when unsupported formatting is used.
+	 */
+	private function parseToken($token, $args, $locale)
+	{
+		$param = trim($token[0]);
+		if (isset($args[$param])) {
+			$arg = $args[$param];
+		} else {
+			return '{' . implode(',', $token) . '}';
+		}
+		$type = isset($token[1]) ? trim($token[1]) : 'none';
+		switch($type)
+		{
+			case 'number':
+			case 'date':
+			case 'time':
+			case 'spellout':
+			case 'ordinal':
+			case 'duration':
+			case 'choice':
+			case 'selectordinal':
+				throw new NotSupportedException("Message format '$type' is not supported. You have to install PHP intl extension to use this feature.");
+			case 'none': return $arg;
+			case 'select':
+				/* http://icu-project.org/apiref/icu4c/classicu_1_1SelectFormat.html
+				selectStyle = (selector '{' message '}')+
+				*/
+				$select = static::tokenizePattern($token[2]);
+				$c = count($select);
+				$message = false;
+				for($i = 0; $i + 1 < $c; $i++) {
+					if (is_array($select[$i]) || !is_array($select[$i + 1])) {
+						return false;
+					}
+					$selector = trim($select[$i++]);
+					if ($message === false && $selector == 'other' || $selector == $arg) {
+						$message = implode(',', $select[$i]);
+					}
+				}
+				if ($message !== false) {
+					return $this->fallbackFormat($locale, $message, $args);
+				}
+			break;
+			case 'plural':
+				/* http://icu-project.org/apiref/icu4c/classicu_1_1PluralFormat.html
+				pluralStyle = [offsetValue] (selector '{' message '}')+
+				offsetValue = "offset:" number
+				selector = explicitValue | keyword
+				explicitValue = '=' number  // adjacent, no white space in between
+				keyword = [^[[:Pattern_Syntax:][:Pattern_White_Space:]]]+
+				message: see MessageFormat
+				*/
+				$plural = static::tokenizePattern($token[2]);
+				$c = count($plural);
+				$message = false;
+				$offset = 0;
+				for($i = 0; $i + 1 < $c; $i++) {
+					if (is_array($plural[$i]) || !is_array($plural[$i + 1])) {
+						return false;
+					}
+					$selector = trim($plural[$i++]);
+					if ($i == 1 && substr($selector, 0, 7) == 'offset:') {
+						$offset = (int) trim(mb_substr($selector, 7, ($pos = mb_strpos(str_replace(["\n", "\r", "\t"], ' ', $selector), ' ', 7)) - 7));
+						$selector = trim(mb_substr($selector, $pos + 1));
+					}
+					if ($message === false && $selector == 'other' ||
+						$selector[0] == '=' && (int) mb_substr($selector, 1) == $arg ||
+						$selector == 'zero' && $arg - $offset == 0 ||
+						$selector == 'one' && $arg - $offset == 1 ||
+						$selector == 'two' && $arg - $offset == 2
+					) {
+						$message = implode(',', str_replace('#', $arg - $offset, $plural[$i]));
+					}
+				}
+				if ($message !== false) {
+					return $this->fallbackFormat($locale, $message, $args);
+				}
+				break;
+		}
+		return false;
 	}
 }
